chore: sanitize codebase, fix lint/type-check/test failures
- apps/web: split AuthContext and DarkModeContext value/hook into
  separate files so react-refresh/only-export-components no longer
  warns. Lint, type-check, build, and vitest all pass with
  --max-warnings 0.
- data-pipeline + scripts: cleared 28 ruff findings (16 unused imports,
  6 unused local vars, 5 stray f-string prefixes, 1 SQLAlchemy
  `== True` → `.is_(True)`). All ruff checks pass repo-wide.
