+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **Sanitized codebase to fix lint/type-check failures**: split React Context constants and the `useDarkMode` hook into their own files (`AuthContextValue.ts`, `DarkModeContextValue.ts`, `useDarkMode.ts`) so `react-refresh/only-export-components` is no longer warned. Cleared 28 ruff findings across `data-pipeline/` and `scripts/` (16 unused imports, 6 unused locals, 5 stray f-string prefixes, 1 SQLAlchemy `== True` → `.is_(True)`). `npm run lint -- --max-warnings 0`, `npm run type-check`, `npm run build`, `npm test`, and `ruff check .` all pass.
+
 ### Fixed
 
 - **Aquafit filter on /schedule only showed Norseman pool**: the two ingestion parsers were tagging the same activity differently — `data-pipeline/sources/toronto_drop_in_api.py` used `AQUAFIT` while `data-pipeline/sources/toronto_parks_json_api.py` and the frontend `SwimType` enum used `AQUATIC_FITNESS`. The drop-in parser now also writes `AQUATIC_FITNESS`, so aquafit sessions from every indoor pool surface under the "Aquatic Fitness" filter button. Existing rows can be relabeled with `UPDATE sessions SET swim_type = 'AQUATIC_FITNESS' WHERE swim_type = 'AQUAFIT';` (no-op on prod where the count is currently 0, but kept for completeness).
