+++ apps/web/src/contexts/DarkModeContextValue.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export type DarkModeContextType = {
+  isDarkMode: boolean;
+  toggleDarkMode: () => void;
+};
+
+export const DarkModeContext = createContext<DarkModeContextType | undefined>(
+  undefined,
+);
