+++ apps/web/src/contexts/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "./AuthContext";
+import { AuthContext } from "./AuthContextValue";
 
 export function useAuth() {
   const context = useContext(AuthContext);
