+++ data-pipeline/config.py
@@ -1,6 +1,5 @@
 """Configuration for data pipeline."""
 from pydantic_settings import BaseSettings
-from typing import Optional
 
 
 class Settings(BaseSettings):
