+++ data-pipeline/jobs/discover_facility_urls.py
@@ -9,7 +9,6 @@ This job:
 """
 import sys
 from pathlib import Path
-from urllib.parse import quote
 
 sys.path.insert(0, str(Path(__file__).parent.parent))
